requirements.txt
matplotlib-inline==0.1.6
networkx==3.3
numpy==1.26.4
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.4
einops==0.8.0
trl==0.13.0
typing_extensions==4.12.2
wandb==0.19.2
zanj==0.5.0
pyswip
maze-dataset